pyproject.toml
[tool.poetry]
name = "monitoring-fastapi"
version = "0.1.0"
description = ""
authors = ["pli <chordplaylist@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
pydantic = "^2.8.2"
opentelemetry-api = "^1.26.0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-distro = "^0.47b0"
prometheus-fastapi-instrumentator = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
